Rename file binarySearch.ts
