+++ src/interpreter/binarySearch.ts
